+++ src/Exception/ServerMessageThrower.php
@@ -106,6 +106,7 @@ class ServerMessageThrower
             'invalid_user', // error_type
         ],
         'Checkpoint\ChangePasswordException'                 => ['/reset(.*?)password/', 'set_new_password', 'change_password'],
+        'Checkpoint\DummyStepException'                      => ['dummy_step'],
         'Checkpoint\RecaptchaChallengeException'             => ['RecaptchaChallengeForm', 'RecaptchaRestrictChallengeForm'],
         'Checkpoint\EscalationChallengeInformationException' => [
             'EscalationChallengeInformationalForm',